refactor(groups): split 1954-line contract router into a folder
Splits the monolithic groupsContractRouter.ts into a groupsContract/ folder:
shared helpers/state, plus discovery / core / content handler modules composed
in index.ts. Each handler binds to its contract route via s.route(...), so the
spread into s.router(...) preserves full per-route type inference — the split
is as type-safe as the original inline router.
